Adapt Santander and Openbank scrappers to project's structure
