+++ lib/bank_scrap.rb
@@ -13,9 +13,9 @@ require 'bank_scrap/banks/banks'
 module BankScrap
   # autoload only requires the file when the specified
   # constant is used for the first time
-  autoload :Bankinter,       'bank_scrap/banks/bankinter'
-  autoload :Bbva,            'bank_scrap/banks/bbva'
-  autoload :Ing,             'bank_scrap/banks/ing'
-  Banks.autoload :Openbank,  'bank_scrap/banks/openbank/openbank'
-  Banks.autoload :Santander, 'bank_scrap/banks/santander/santander'
+  autoload :Bankinter, 'bank_scrap/banks/bankinter'
+  autoload :Bbva,      'bank_scrap/banks/bbva'
+  autoload :Ing,       'bank_scrap/banks/ing'
+  autoload :Openbank,  'bank_scrap/banks/openbank'
+  autoload :Santander, 'bank_scrap/banks/santander'
 end
