+++ lib/bank_scrap/accounts/openbank_account.rb
@@ -1,0 +1,5 @@
+module BankScrap
+  class OpenbankAccount < Account
+    attr_accessor :contract_id
+  end
+end
