+++ lib/bank_scrap/accounts/santander_account.rb
@@ -1,0 +1,5 @@
+module BankScrap
+  class SantanderAccount < Account
+    attr_accessor :contract_id
+  end
+end
