--- lib/bank_scrap/banks/openbank/account.rb
@@ -1,7 +1,0 @@
-module BankScrap::Banks
-  module Openbank
-    class Account < BankScrap::Account
-      attr_accessor :contract_id
-    end
-  end
-end
