--- lib/bank_scrap/banks/santander/account.rb
@@ -1,7 +1,0 @@
-module BankScrap::Banks
-  module Santander
-    class Account < BankScrap::Account
-      attr_accessor :contract_id
-    end
-  end
-end
