--- lib/bank_scrap/banks/santander/santander.rb
@@ -1,7 +1,0 @@
-module BankScrap
-  module Banks
-    module Santander
-      BankScrap::Banks::Santander.autoload :Bank,   'bank_scrap/banks/santander/bank'
-    end
-  end
-end
