+++ lib/bank_scrap/cli.rb
@@ -77,7 +77,7 @@ module BankScrap
     end
 
     def find_bank_class_for(bank_name)
-      Object.const_get("BankScrap::#{bank_name}".classify) rescue Object.const_get("BankScrap::Banks::#{bank_name}::Bank".classify)
+      Object.const_get("BankScrap::#{bank_name}".classify)
     rescue NameError
       raise ArgumentError.new('Invalid bank name')
     end
